Added Prepayment and settlement, refactored code

+++ accruals-accounting-service/src/main/java/com/ebooks/accrualsaccountingservice/services/InterestInfoService.java
@@ -1,11 +1,11 @@
 package com.ebooks.accrualsaccountingservice.services;
 
-import com.ebooks.accrualsaccountingservice.entities.EMISchedule;
-import com.ebooks.accrualsaccountingservice.entities.InterestInfo;
+
+import com.ebooks.commonmoduleloan.entities.InterestInfo;
 
 import java.time.LocalDate;
 
 public interface InterestInfoService {
     void createAccrual(InterestInfo interestInfo);
-    boolean existsByLoanAndInstallmentAndDate(String loanNumber, Long installmentNumber, LocalDate today);
+    boolean existsByLoanAndInstallmentAndDate(String loanNumber, Integer installmentNumber, LocalDate today);
 }
